+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 05/02/2015
+## 05/02/2015 - v0.6.0
+- Convert from `Coffeescript` to `ES6 (Babel)`
 - Remove `gulp` in favor of `broccoli`
 - Fix `*.json` files to include `main`
 - Restructure directory layout
+- Include bundled (with Tether) and standalone distributions